+++ src/Migration/Mapping/Lookup/SalesChannelLookup.php
@@ -1,0 +1,62 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace SwagMigrationAssistant\Migration\Mapping\Lookup;
+
+use Shopware\Core\Framework\Context;
+use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\MultiFilter;
+use Shopware\Core\Framework\Log\Package;
+use Shopware\Core\System\SalesChannel\SalesChannelCollection;
+use Symfony\Contracts\Service\ResetInterface;
+
+#[Package('fundamentals@after-sales')]
+class SalesChannelLookup implements ResetInterface
+{
+    /**
+     * @var array<string, string|null>
+     */
+    private array $cache = [];
+
+    /**
+     * @param EntityRepository<SalesChannelCollection> $salesChannelRepository
+     */
+    public function __construct(
+        private readonly EntityRepository $salesChannelRepository,
+    ) {
+    }
+
+    public function getSalesChannelWithTypeAndName(string $typeId, string $name, Context $context): ?string
+    {
+        $cacheKey = $typeId . '-' . $name;
+
+        if (\array_key_exists($cacheKey, $this->cache)) {
+            return $this->cache[$cacheKey];
+        }
+
+        $criteria = new Criteria();
+        $criteria->addFilter(new MultiFilter(
+            MultiFilter::CONNECTION_AND,
+            [
+                new EqualsFilter('typeId', $typeId),
+                new EqualsFilter('translations.name', $name),
+            ]
+        ));
+        $criteria->setLimit(1);
+
+        $this->cache[$cacheKey] = $this->salesChannelRepository->searchIds($criteria, $context)->firstId();
+
+        return $this->cache[$cacheKey];
+    }
+
+    public function reset(): void
+    {
+        $this->cache = [];
+    }
+}