+++ src/Profile/Shopware6/Converter/OrderConverter.php
@@ -10,6 +10,8 @@ namespace SwagMigrationAssistant\Profile\Shopware6\Converter;
 use Shopware\Core\Framework\Log\Package;
 use SwagMigrationAssistant\Migration\Converter\ConvertStruct;
 use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
+use SwagMigrationAssistant\Migration\Logging\Log\Builder\MigrationLogBuilder;
+use SwagMigrationAssistant\Migration\Logging\Log\ConvertAssociationMissingLog;
 use SwagMigrationAssistant\Migration\Logging\LoggingServiceInterface;
 use SwagMigrationAssistant\Migration\Mapping\Lookup\StateMachineStateLookup;
 use SwagMigrationAssistant\Migration\Mapping\MappingServiceInterface;
@@ -58,6 +60,18 @@ class OrderConverter extends ShopwareConverter
             DefaultEntities::SALES_CHANNEL,
             $converted['salesChannelId']
         );
+
+        if ($converted['salesChannelId'] === null) {
+            $this->loggingService->log(
+                MigrationLogBuilder::fromMigrationContext($this->migrationContext)
+                    ->withEntityName(DefaultEntities::ORDER)
+                    ->withFieldName('salesChannelId')
+                    ->withSourceData($data)
+                    ->build(ConvertAssociationMissingLog::class)
+            );
+
+            return new ConvertStruct(null, $data);
+        }
 
         $converted['orderCustomer']['salutationId'] = $this->getMappingIdFacade(
             DefaultEntities::SALUTATION,