+++ src/Profile/Shopware6/Converter/PageSystemConfigConverter.php
@@ -10,6 +10,8 @@ namespace SwagMigrationAssistant\Profile\Shopware6\Converter;
 use Shopware\Core\Framework\Log\Package;
 use SwagMigrationAssistant\Migration\Converter\ConvertStruct;
 use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
+use SwagMigrationAssistant\Migration\Logging\Log\Builder\MigrationLogBuilder;
+use SwagMigrationAssistant\Migration\Logging\Log\ConvertAssociationMissingLog;
 use SwagMigrationAssistant\Migration\Logging\LoggingServiceInterface;
 use SwagMigrationAssistant\Migration\Mapping\Lookup\SystemConfigLookup;
 use SwagMigrationAssistant\Migration\Mapping\MappingServiceInterface;
@@ -38,7 +40,27 @@ class PageSystemConfigConverter extends ShopwareConverter
     {
         $converted = $data;
 
-        $systemConfigUuid = $this->systemConfigLookup->get($data['configurationKey'], $data['salesChannelId'] ?? null, $this->context);
+        $mappedSalesChannelId = null;
+        if (isset($data['salesChannelId'])) {
+            $mappedSalesChannelId = $this->getMappingIdFacade(
+                DefaultEntities::SALES_CHANNEL,
+                $data['salesChannelId']
+            );
+
+            if ($mappedSalesChannelId === null) {
+                $this->loggingService->log(
+                    MigrationLogBuilder::fromMigrationContext($this->migrationContext)
+                        ->withEntityName(DefaultEntities::PAGE_SYSTEM_CONFIG)
+                        ->withFieldName('salesChannelId')
+                        ->withSourceData($data)
+                        ->build(ConvertAssociationMissingLog::class)
+                );
+
+                return new ConvertStruct(null, $data);
+            }
+        }
+
+        $systemConfigUuid = $this->systemConfigLookup->get($data['configurationKey'], $mappedSalesChannelId, $this->context);
         if ($systemConfigUuid !== null) {
             $converted['id'] = $systemConfigUuid;
         }
@@ -50,10 +72,7 @@ class PageSystemConfigConverter extends ShopwareConverter
         );
 
         if (isset($converted['salesChannelId'])) {
-            $converted['salesChannelId'] = $this->getMappingIdFacade(
-                DefaultEntities::SALES_CHANNEL,
-                $converted['salesChannelId']
-            );
+            $converted['salesChannelId'] = $mappedSalesChannelId;
         }
 
         return new ConvertStruct($converted, null, $this->mainMapping['id'] ?? null);