+++ src/Profile/Shopware6/Converter/SalesChannelDomainConverter.php
@@ -10,6 +10,8 @@ namespace SwagMigrationAssistant\Profile\Shopware6\Converter;
 use Shopware\Core\Framework\Log\Package;
 use SwagMigrationAssistant\Migration\Converter\ConvertStruct;
 use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
+use SwagMigrationAssistant\Migration\Logging\Log\Builder\MigrationLogBuilder;
+use SwagMigrationAssistant\Migration\Logging\Log\ConvertAssociationMissingLog;
 use SwagMigrationAssistant\Migration\MigrationContextInterface;
 use SwagMigrationAssistant\Profile\Shopware6\DataSelection\DataSet\SalesChannelDomainDataSet;
 use SwagMigrationAssistant\Profile\Shopware6\Shopware6MajorProfile;
@@ -38,10 +40,28 @@ class SalesChannelDomainConverter extends ShopwareConverter
         $converted['snippetSetId'] = $this->getMappingIdFacade(DefaultEntities::SNIPPET_SET, $data['snippetSetId']);
         $converted['salesChannelId'] = $this->getMappingIdFacade(DefaultEntities::SALES_CHANNEL, $data['salesChannelId']);
 
+        if ($converted['salesChannelId'] === null) {
+            $this->loggingService->log(
+                MigrationLogBuilder::fromMigrationContext($this->migrationContext)
+                    ->withEntityName(DefaultEntities::SALES_CHANNEL_DOMAIN)
+                    ->withFieldName('salesChannelId')
+                    ->withSourceData($data)
+                    ->build(ConvertAssociationMissingLog::class)
+            );
+
+            return new ConvertStruct(null, $data);
+        }
+
         if (isset($data['salesChannelDefaultHreflang'])) {
-            $converted['salesChannelDefaultHreflang'] = [
-                'id' => $this->getMappingIdFacade(DefaultEntities::SALES_CHANNEL, $data['salesChannelDefaultHreflang']['id']),
-            ];
+            $salesChannelDefaultHreflangId = $this->getMappingIdFacade(DefaultEntities::SALES_CHANNEL, $data['salesChannelDefaultHreflang']['id']);
+
+            if ($salesChannelDefaultHreflangId !== null) {
+                $converted['salesChannelDefaultHreflang'] = [
+                    'id' => $salesChannelDefaultHreflangId,
+                ];
+            } else {
+                unset($converted['salesChannelDefaultHreflang']);
+            }
         }
 
         return new ConvertStruct($converted, null, $this->mainMapping['id'] ?? null);