+++ src/Profile/Shopware6/Converter/SeoUrlTemplateConverter.php
@@ -10,6 +10,8 @@ namespace SwagMigrationAssistant\Profile\Shopware6\Converter;
 use Shopware\Core\Framework\Log\Package;
 use SwagMigrationAssistant\Migration\Converter\ConvertStruct;
 use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
+use SwagMigrationAssistant\Migration\Logging\Log\Builder\MigrationLogBuilder;
+use SwagMigrationAssistant\Migration\Logging\Log\ConvertAssociationMissingLog;
 use SwagMigrationAssistant\Migration\Logging\LoggingServiceInterface;
 use SwagMigrationAssistant\Migration\Mapping\Lookup\SeoUrlTemplateLookup;
 use SwagMigrationAssistant\Migration\Mapping\MappingServiceInterface;
@@ -38,8 +40,28 @@ class SeoUrlTemplateConverter extends ShopwareConverter
     {
         $converted = $data;
 
+        $mappedSalesChannelId = null;
+        if (isset($data['salesChannelId'])) {
+            $mappedSalesChannelId = $this->getMappingIdFacade(
+                DefaultEntities::SALES_CHANNEL,
+                $data['salesChannelId']
+            );
+
+            if ($mappedSalesChannelId === null) {
+                $this->loggingService->log(
+                    MigrationLogBuilder::fromMigrationContext($this->migrationContext)
+                        ->withEntityName(DefaultEntities::SEO_URL_TEMPLATE)
+                        ->withFieldName('salesChannelId')
+                        ->withSourceData($data)
+                        ->build(ConvertAssociationMissingLog::class)
+                );
+
+                return new ConvertStruct(null, $data);
+            }
+        }
+
         $seoUrlTemplateUuid = $this->seoUrlTemplateLookup->get(
-            $data['salesChannelId'] ?? null,
+            $mappedSalesChannelId,
             $data['routeName'],
             $this->context
         );
@@ -55,10 +77,7 @@ class SeoUrlTemplateConverter extends ShopwareConverter
         );
 
         if (isset($converted['salesChannelId'])) {
-            $converted['salesChannelId'] = $this->getMappingIdFacade(
-                DefaultEntities::SALES_CHANNEL,
-                $converted['salesChannelId']
-            );
+            $converted['salesChannelId'] = $mappedSalesChannelId;
         }
 
         return new ConvertStruct($converted, null, $this->mainMapping['id'] ?? null);