+++ tests/_fixtures/Shopware6/PageSystemConfig/01-NotExists/output.php
@@ -5,9 +5,4 @@
  * file that was distributed with this source code.
  */
 
-return [
-    'configurationKey' => 'core.basicInformation.contactPage',
-    'configurationValue' => 'b0dfd6d77eea4cbaa02b35e8562d823d',
-    'salesChannelId' => null,
-    'id' => 'a4b495f06c1e49279dea600ffea12794',
-];
+return null;