+++ tests/_fixtures/Shopware6/Product/10-DropUnsupportedSalesChannelVisibilities/mapping.php
@@ -1,0 +1,36 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
+
+return [
+    [
+        'entityName' => DefaultEntities::TAX,
+        'oldIdentifier' => 'a2f2918aaf5d42a2ba219c82c2bc275d',
+        'newIdentifier' => 'a2f2918aaf5d42a2ba219c82c2bc275d',
+    ],
+    [
+        'entityName' => DefaultEntities::PRODUCT_MANUFACTURER,
+        'oldIdentifier' => '954217f3551b4fe091d9cdea5731d582',
+        'newIdentifier' => '954217f3551b4fe091d9cdea5731d582',
+    ],
+    [
+        'entityName' => DefaultEntities::CURRENCY,
+        'oldIdentifier' => 'b7d2554b0ce847cd82f3ac9bd1c0dfca',
+        'newIdentifier' => 'b7d2554b0ce847cd82f3ac9bd1c0dfca',
+    ],
+    [
+        'entityName' => DefaultEntities::UNIT,
+        'oldIdentifier' => '284217f3551b4fe091d9cdea5731d582',
+        'newIdentifier' => '284217f3551b4fe091d9cdea5731d582',
+    ],
+    [
+        'entityName' => DefaultEntities::SALES_CHANNEL,
+        'oldIdentifier' => '3f05f1f6514d43f7ae11a669c7557d1c',
+        'newIdentifier' => 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+    ],
+];