+++ tests/_fixtures/Shopware6/SalesChannel/04-UnsupportedType/input.php
@@ -1,0 +1,21 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+return [
+    'typeId' => 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+    'languageId' => '24f78d24f14849f6922ab2f76dcd76e8',
+    'currencyId' => 'b7d2554b0ce847cd82f3ac9bd1c0dfca',
+    'paymentMethodId' => 'c5917da8076b495ba80b14a61afd90fb',
+    'shippingMethodId' => '6e028d0f24114544ad25235d56bb8846',
+    'countryId' => '580a3bdd1739487f99ddd56ec365828f',
+    'navigationCategoryId' => '7515fecfbfe14fbb87892873dbd1134d',
+    'navigationCategoryDepth' => 2,
+    'name' => 'Social Shopping',
+    'accessKey' => 'SWSCA83PBVT8UA_X_P5UF6IJKG',
+    'translations' => [],
+    'id' => '99999999999999999999999999999999',
+];